[chore] refactor API tests
