+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start:dev": "nodemon --exec babel-node ./server/bin/www",
     "start": "babel-node ./server/bin/www",
-    "test": "NODE_ENV=test sequelize db:migrate && NODE_ENV=test nyc mocha --compilers js:babel-core/register server/test --timeout 50000",
+    "test": "NODE_ENV=test sequelize db:migrate && NODE_ENV=test nyc mocha --compilers js:babel-core/register server/test/userController server/test/groupController server/test/messageController server/test/helpers --timeout 50000",
     "lint": "esw server --color",
     "build-css": "node-sass --include-path scss ./template/styles/main.scss ./template/styles/style.css",
     "watch-css": "nodemon -e scss -x \"npm run build-css\"",
@@ -105,6 +105,7 @@
     "sequelize-cli": "^2.7.0",
     "shortid": "^2.2.8",
     "style-loader": "^0.18.2",
+    "supertest": "^3.0.0",
     "toastr": "^2.1.2",
     "webpack": "^3.3.0",
     "webpack-dev-middleware": "^1.11.0",
