--- server/test/tests.js
@@ -1,912 +1,0 @@
-// API Tests
-
-import chaiHttp from 'chai-http';
-import chai from 'chai';
-import app from '../app';
-import models from '../models';
-import customSort from '../../helpers/customSort';
-
-process.env.NODE_ENV = 'test';
-const should = chai.should();
-chai.use(chaiHttp);
-let token;
-
-models.User.destroy({
-  where: {},
-  cascade: true,
-  truncate: true,
-  restartIdentity: true
-});
-models.Message.destroy({
-  where: {},
-  cascade: true,
-  truncate: true,
-  restartIdentity: true
-});
-models.Group.destroy({
-  where: {},
-  cascade: true,
-  truncate: true,
-  restartIdentity: true
-});
-models.UserGroup.destroy({
-  where: {},
-  cascade: true,
-  truncate: true,
-  restartIdentity: true
-});
-
-describe('PostIT API Tests:', () => {
-  describe('Base API', () => {
-    it('Displays welcome message', (done) => {
-      chai.request(app)
-        .get('/api')
-        .type('form')
-        .end((err, res) => {
-          res.body.message.should.equals('Welcome to the Post IT API!');
-          done();
-        });
-    });
-  });
-  describe('Signup API route', () => {
-    it('Registers a new user when complete parameters are provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/user/signup')
-        .type('form')
-        .send({
-          username: 'chuks',
-          email: 'chuks@andela.com',
-          password: 'chukspass',
-          phone: '2347033130448'
-        })
-        .end((err, res) => {
-          res.should.have.status(201);
-          done();
-        });
-    });
-    it('Returns 400 error and error message with no username parameter',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            email: 'chuks@andela.com',
-            password: 'chukspass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.username.should.equals(
-              'Username field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with no email parameter',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'chuks',
-            password: 'chukspass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.email.should.equals(
-              'Email address field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with no password parameter',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'chuks',
-            email: 'chuks@andela.com',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.password.should.equals(
-              'Password field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with no phone parameter',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'chuks',
-            email: 'chuks@andela.com',
-            password: 'chukspass'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.phone.should.equals('Phone field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with a duplicate username',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'chuks',
-            email: 'chukss@andela.com',
-            password: 'chukspass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.username.should.equals('Username already exists');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with a duplicate email address',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'dave',
-            email: 'chuks@andela.com',
-            password: 'chukspass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.email.should.equals('Email address already exists');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with an invalid email address',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'funsho',
-            email: 'funshoandela.com',
-            password: 'funshopass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.email.should.equals('Email address is invalid');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with an empty username field',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: '  ',
-            email: 'dave@andela.com',
-            password: 'davepass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.username.should.equals(
-              'Username field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with an empty email field',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'dave',
-            email: '   ',
-            password: 'davepass',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.email.should.equals(
-              'Email address field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with an empty password field',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'dave',
-            email: 'dave@andela.com',
-            password: '   ',
-            phone: '2347033130448'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.password.should.equals(
-              'Password field cannot be empty');
-            done();
-          });
-      });
-    it('Returns 400 error and error message with an empty phone field',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signup')
-          .type('form')
-          .send({
-            username: 'dave',
-            email: 'dave@andela.com',
-            password: 'davepass',
-            phone: '  '
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.errors.phone.should.equals('Phone field cannot be empty');
-            done();
-          });
-      });
-  });
-  describe('Signin API route', () => {
-    it('Successful signin should return 200 status code', (done) => {
-      chai.request(app)
-        .post('/api/v1/user/signin')
-        .type('form')
-        .send({
-          username: 'chuks',
-          password: 'chukspass',
-        })
-        .end((err, res) => {
-          res.should.have.status(200);
-          done();
-        });
-    });
-    it('Successful signin should return a token', (done) => {
-      chai.request(app)
-        .post('/api/v1/user/signin')
-        .type('form')
-        .send({
-          username: 'chuks',
-          password: 'chukspass',
-        })
-        .end((err, res) => {
-          res.body.data.should.have.property('token');
-          token = res.body.data.token;
-          done();
-        });
-    });
-    it('Returns error message when no username is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/user/signin')
-        .type('form')
-        .send({
-          password: 'chukspass',
-        })
-        .end((err, res) => {
-          res.body.errors.username.should.equals(
-            'Username field cannot be empty');
-          done();
-        });
-    });
-    it('Returns error message when no password is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/user/signin')
-        .type('form')
-        .send({
-          username: 'chuks',
-        })
-        .end((err, res) => {
-          res.body.errors.password.should.equals(
-            'Password field cannot be empty');
-          done();
-        });
-    });
-    it('Returns error message when none existing username is provided',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/user/signin')
-          .type('form')
-          .send({
-            username: 'chukss',
-            password: 'chukspass'
-          })
-          .end((err, res) => {
-            res.body.errors.username.should.equals('User does not exist');
-            done();
-          });
-      });
-    it('Returns error message when wrong password is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/user/signin')
-        .type('form')
-        .send({
-          username: 'chuks',
-          password: 'chukspasss'
-        })
-        .end((err, res) => {
-          res.body.errors.password.should.equals('Incorrect password!');
-          done();
-        });
-    });
-  });
-  describe('API route to find all Users', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .get('/api/v1/users')
-        .type('form')
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns an array of user objects when token is valid', (done) => {
-      chai.request(app)
-        .get('/api/v1/users/')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(200);
-          res.body.data.user.should.be.an('array');
-          res.body.data.user[0].should.be.an('object');
-          done();
-        });
-    });
-  });
-  describe('API route to fetch logged in user group info', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .get('/api/v1/users/one')
-        .type('form')
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    // it('returns logged in users group info when token is valid', (done) => {
-    //   chai.request(app)
-    //     .get('/api/v1/users/one')
-    //     .type('form')
-    //     .set('x-access-token', token)
-    //     .end((err, res) => {
-    //       res.should.have.status(200);
-    //       // res.body.data.should.be.an('object');
-    //       // res.body.data[0].should.have.property('unreadCount');
-    //       done();
-    //     });
-    // });
-  });
-  describe('Search user API route', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .get('/api/v1/user/c/0/2/find')
-        .type('form')
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns logged in user object with group info when token is valid',
-      (done) => {
-        chai.request(app)
-          .get('/api/v1/user/chuks/0/0/find')
-          .type('form')
-          .set('x-access-token', token)
-          .end((err, res) => {
-            res.should.have.status(200);
-            res.body.data.count.should.equals(1);
-            res.body.data.should.be.an('object');
-            done();
-          });
-      });
-  });
-  describe('Create group API route', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group')
-        .type('form')
-        .send({
-          name: 'Andela Bootcamp',
-          description: 'A little group description'
-        })
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns error if no parameter is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(400);
-          res.body.errors.name.should.equals('Please choose a group name');
-          res.body.errors.description.should.equals(
-            'Please enter a description of the group');
-          done();
-        });
-    });
-    it('creates a group if all parameters are provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          name: 'Andela Bootcamp',
-          description: 'A little group description'
-        })
-        .end((err, res) => {
-          res.should.have.status(201);
-          res.body.data.group.name.should.equals('Andela Bootcamp');
-          res.body.data.group.description.should.equals(
-            'A little group description');
-          done();
-        });
-    });
-    it('returns error if provided group name already exists', (done) => {
-      chai.request(app)
-        .post('/api/v1/group')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          name: 'Andela Bootcamp',
-          description: 'A little group description'
-        })
-        .end((err, res) => {
-          res.should.have.status(400);
-          res.body.errors.group.should.equals('Group already exists');
-          done();
-        });
-    });
-  });
-  describe('API route for adding users to a group', () => {
-    before((done) => {
-      chai.request(app)
-        .post('/api/v1/user/signup')
-        .type('form')
-        .send({
-          username: 'dave',
-          email: 'dave@andela.com',
-          password: 'davepass',
-          phone: '2347033130449'
-        })
-        .end(() => {
-          done();
-        });
-    });
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/user')
-        .type('form')
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns error if no user id is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/user')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(400);
-          res.body.error.message.should.equals('a User ID is required');
-          done();
-        });
-    });
-    it('returns error if user id is provided does not exist', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/user')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          userId: 8
-        })
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('User does not exist');
-          done();
-        });
-    });
-    it('returns error if group id is provided does not exist', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/7/user')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          userId: 2
-        })
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('Group does not exist');
-          done();
-        });
-    });
-    it('returns error if user id is provided already belongs to the group',
-      (done) => {
-        chai.request(app)
-          .post('/api/v1/group/1/user')
-          .type('form')
-          .set('x-access-token', token)
-          .send({
-            userId: 1
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.error.message.should.equals(
-              'User already belongs to this group');
-            done();
-          });
-      });
-    it('adds the user to the group', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/user')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          userId: 4
-        })
-        .end((err, res) => {
-          res.should.have.status(201);
-          res.body.data.usergroup.userId.should.equals(4);
-          res.body.data.usergroup.groupId.should.equals(1);
-          done();
-        });
-    });
-  });
-  describe('API route for posting message to a particular group', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/message')
-        .type('form')
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns error message if message title is not provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/message')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(400);
-          res.body.error.message.should.equals(
-            'Message title can not be empty');
-          done();
-        });
-    });
-    it('returns error message if message body is not provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/message')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          title: 'A message title'
-        })
-        .end((err, res) => {
-          res.should.have.status(400);
-          res.body.error.message.should.equals('Message can not be empty');
-          done();
-        });
-    });
-    it('returns error message if group id provided does not exist', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/44/message')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          title: 'A message title',
-          message: 'a message body'
-        })
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('That group does not exist');
-          done();
-        });
-    });
-    it('sets message priority to normal if it is not provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/message')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          title: 'A message title',
-          message: 'a message body'
-        })
-        .end((err, res) => {
-          res.should.have.status(201);
-          res.body.data.message.priority.should.equals('normal');
-          done();
-        });
-    });
-    it('sends message if all parameters are provided', (done) => {
-      chai.request(app)
-        .post('/api/v1/group/1/message')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          title: 'A message title',
-          message: 'a message body',
-          priority: 'critical'
-        })
-        .end((err, res) => {
-          res.should.have.status(201);
-          res.body.data.message.priority.should.equals('critical');
-          res.body.data.message.title.should.equals('A message title');
-          res.body.data.message.message.should.equals('a message body');
-          res.body.data.message.author.should.equals('chuks');
-          done();
-        });
-    });
-  });
-  describe('API route for feching messages', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .get('/api/v1/group/1/messages')
-        .type('form')
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns error message if group id provided does not exist', (done) => {
-      chai.request(app)
-        .get('/api/v1/group/44/messages')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('Group does not exist');
-          done();
-        });
-    });
-    it('returns an array of messages if all parameters are correct', (done) => {
-      chai.request(app)
-        .get('/api/v1/group/1/messages')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(200);
-          res.body.data.should.be.an('array');
-          res.body.data[0].should.be.an('object');
-          done();
-        });
-    });
-  });
-  describe('Reset password API route', () => {
-    it('returns 404 error with an error message if no email is provided',
-      (done) => {
-        chai.request(app)
-          .patch('/api/v1/user/reset')
-          .type('form')
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.error.should.equals('No email address provided');
-            done();
-          });
-      });
-    it('returns 404 error with an error message if no request type is provided',
-      (done) => {
-        chai.request(app)
-          .patch('/api/v1/user/reset')
-          .type('form')
-          .send({
-            email: 'chuks@andela.com'
-          })
-          .end((err, res) => {
-            res.should.have.status(400);
-            res.body.error.should.equals('Request type must be specified');
-            done();
-          });
-      });
-    // it('returns 404 error with an error message if invalid request type is provided',(done) => {
-    //   chai.request(app)
-    //     .patch('/api/v1/user/reset')
-    //     .type('form')
-    //     .send({
-    //       email: 'chuks@andela.com',
-    //       type: 'break it'
-    //     })
-    //     .end((err, res) => {
-    //       res.should.have.status(400);
-    //       res.body.error.should.equals('Valid request type must be specified');
-    //       done();
-    //     });
-    // });
-  });
-  describe('API route for updating message readby', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .patch('/api/v1/message/readby')
-        .type('form')
-        .send({
-          id: 1,
-          readby: ['chuks', 'tayo']
-        })
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns error if user has read the message', (done) => {
-      chai.request(app)
-        .patch('/api/v1/message/readby')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          id: 1,
-          readby: 'chuks'
-        })
-        .end((err, res) => {
-          res.should.have.status(400);
-          res.body.errors.should.equal('User has read this message');
-          done();
-        });
-    });
-    it('returns error if user has read the message', (done) => {
-      chai.request(app)
-        .patch('/api/v1/message/readby')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          id: 1,
-          readby: 'jude'
-        })
-        .end((err, res) => {
-          res.should.have.status(200);
-          done();
-        });
-    });
-  });
-  describe('API route for removing users from a group', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .patch('/api/v1/group/1/remove')
-        .type('form')
-        .send({
-          userId: 1,
-        })
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns an error if no user id is provided', (done) => {
-      chai.request(app)
-        .patch('/api/v1/group/1/remove')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-        })
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('User and group id must be provided');
-          done();
-        });
-    });
-    it('returns an error if no non-existing group id is provided', (done) => {
-      chai.request(app)
-        .patch('/api/v1/group/111/remove')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          userId: 1
-        })
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('User or group does not exist');
-          done();
-        });
-    });
-    it('returns an error if no non-existing user id is provided', (done) => {
-      chai.request(app)
-        .patch('/api/v1/group/1/remove')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          userId: 111
-        })
-        .end((err, res) => {
-          res.should.have.status(401);
-          res.body.error.message.should.equals('User or group does not exist');
-          done();
-        });
-    });
-    it('removes a user when correct details are provided', (done) => {
-      chai.request(app)
-        .patch('/api/v1/group/1/remove')
-        .type('form')
-        .set('x-access-token', token)
-        .send({
-          userId: 1
-        })
-        .end((err, res) => {
-          res.should.have.status(200);
-          res.body.removedUser.should.equals(1);
-          done();
-        });
-    });
-  });
-  describe('API route for getting group member count', () => {
-    it('returns error if no token is provided', (done) => {
-      chai.request(app)
-        .get('/api/v1/group/1/count')
-        .type('form')
-        .send({
-          userId: 1,
-        })
-        .end((err, res) => {
-          res.should.have.status(403);
-          res.body.message.should.equals(
-            'User not authenticated. Failed to authenticate token.');
-          done();
-        });
-    });
-    it('returns the member count of a group', (done) => {
-      chai.request(app)
-        .get('/api/v1/group/1/count')
-        .type('form')
-        .set('x-access-token', token)
-        .end((err, res) => {
-          res.should.have.status(200);
-          res.body.data.should.equals(2);
-          done();
-        });
-    });
-  });
-  describe('customSort helper function', () => {
-    it('should be a function', (done) => {
-      customSort.should.be.a('function');
-      done();
-    });
-    it('should return an array sorted by ID', (done) => {
-      const array = [
-        {
-          group: {
-            id: 13,
-            name: 'Andela',
-            description: 'an Andelan group'
-          },
-          unreadCount: 1
-        },
-        {
-          group: {
-            id: 2,
-            name: 'PostIT',
-            description: "A group for HNG's Factory product"
-          },
-          unreadCount: 4
-        }
-      ];
-      const sortedArr = array.sort(customSort);
-      sortedArr[0].group.id.should.equal(2);
-      done();
-    });
-  });
-});
